Added most of @nickie's suggestions, and: * added some checks in order for some fail tests to pass. * added a check in `InitContext()` to avoid a segmentation error
When copying from the bytes buffer the following method is used
```
int p = skip(len);
std::copy(bytes.begin() + p, bytes.begin() + pos, &target)
```

This is also the way we read floats and in the case of f64 it
causes a `*** stack smashing detected ***` error (gcc).

I have added the -fno-stack-protector flag, but I am reading that
a production release must not have this flag for security reasons.

**Fix this before merging!**