Add read_name in custom section.
When we read the contents of the custon section a type::Name
is the first thing. After that the is a series of bytes that
we can ignore. The name should be read because it can cause a
module to be invalid/malformed which we care about.

By adding this we pass ~200 more invalid tests!