Fix globals validation
A global.get instruction is considered const when:
* The global it is getting is immutable.
* The global has been imported.

The second condition is used to make sure that we use C' (without
ever creating it) to validate globals. The `is_const()` function
therefore should only be used in that context.

Keep in mind that in order for a global.get to know if it is const,
it first has to validate itself. This is not true with `t.const`
instructions that are trivial.